requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
CacheMan==2.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cyhunspell==2.0.2
futures==3.0.5
goslate==1.5.4
idna==3.4
psutil==5.9.4
PyDictionary==2.0.1
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
